news_collection/lambda_5_finalExport/requirements.txt
boto3==1.35.40
botocore==1.35.40
DateTime==5.5
jmespath==1.0.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.2
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
urllib3==2.2.3
zope.interface==7.1.0
